% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{boostrap_crosscomm_bseds}
\alias{boostrap_crosscomm_bseds}
\title{Bootstrap sample community masses from two communities}
\usage{
boostrap_crosscomm_bseds(community_a, community_b)
}
\arguments{
\item{community_a}{vector of masses for first community}

\item{community_b}{vector of masses for second community}
}
\value{
list of 2 vectors: masses for first resample community, masses for second resample community.
}
\description{
Pool all individual masses from communities and randomly re-draw communities with the original number of individuals from the pool, with replacement.
}
