% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_sampling.R
\name{boostrap_unif_bsed}
\alias{boostrap_unif_bsed}
\title{Bootstrap sample community masses from uniform}
\usage{
boostrap_unif_bsed(nind, min_mass, max_mass)
}
\arguments{
\item{nind}{N}

\item{min_mass}{Min mass}

\item{max_mass}{Max mass}
}
\value{
vector of masses for a community of N individuals
}
\description{
Randomly draw N masses, where N is the total number of individuals in a community, from a uniform distribution with min and max corresponding the min and max of the entire community.
}
