% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doi}
\alias{doi}
\title{Compute distribution overlap index for BSEDs}
\usage{
doi(dist_a, dist_b = NULL)
}
\arguments{
\item{dist_a}{First distribution}

\item{dist_b}{Second distribution, if NULL, uniform}
}
\value{
DOI for two distributions
}
\description{
Calculate DOI for two distributions or one distribution compared to uniform.
}
